+++ src/Command/GenerateCommand.php
@@ -13,8 +13,12 @@ use DoclerLabs\ApiClientGenerator\Output\MetaFilePrinter;
 use DoclerLabs\ApiClientGenerator\Output\Php\PhpFileCollection;
 use DoclerLabs\ApiClientGenerator\Output\PhpFilePrinter;
 use DoclerLabs\ApiClientGenerator\Output\StaticPhpFileCopier;
+use DoclerLabs\ApiClientGenerator\Output\WarningFormatter;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\StyleInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
@@ -23,16 +27,16 @@ use Symfony\Component\Finder\Finder;
 
 class GenerateCommand extends Command
 {
-    private Configuration        $configuration;
-    private CodeGeneratorFacade  $codeGenerator;
-    private FileReader           $fileReader;
-    private Parser               $parser;
-    private PhpFilePrinter       $phpPrinter;
-    private MetaTemplateFacade   $metaTemplate;
-    private MetaFilePrinter      $templatePrinter;
-    private Finder               $fileFinder;
-    private StaticPhpFileCopier  $staticPhpPrinter;
-    private Filesystem           $filesystem;
+    private Configuration       $configuration;
+    private CodeGeneratorFacade $codeGenerator;
+    private FileReader          $fileReader;
+    private Parser              $parser;
+    private PhpFilePrinter      $phpPrinter;
+    private MetaTemplateFacade  $metaTemplate;
+    private MetaFilePrinter     $templatePrinter;
+    private Finder              $fileFinder;
+    private StaticPhpFileCopier $staticPhpPrinter;
+    private Filesystem          $filesystem;
 
     public function __construct(
         Configuration $configuration,
@@ -45,7 +49,8 @@ class GenerateCommand extends Command
         Finder $fileFinder,
         StaticPhpFileCopier $staticPhpCopier,
         Filesystem $filesystem
-    ) {
+    )
+    {
         parent::__construct();
         $this->configuration    = $configuration;
         $this->fileReader       = $fileReader;
@@ -70,6 +75,7 @@ class GenerateCommand extends Command
 
     public function execute(InputInterface $input, OutputInterface $output): int
     {
+        $this->initWarningPrinting($input);
         $specificationFilePath = $this->configuration->getSpecificationFilePath();
 
         $specification = $this->parser->parse(
@@ -161,17 +167,32 @@ class GenerateCommand extends Command
 
     private function copySpecification(StyleInterface $ss): void
     {
-        $ss->text(sprintf('Copy specification file to %s.', $this->configuration->getOutputDirectory()));
-
         $destinationPath = sprintf(
             '%s/doc/%s',
             $this->configuration->getOutputDirectory(),
             basename($this->configuration->getSpecificationFilePath())
         );
 
+        $ss->text(sprintf('Copy specification file to %s.', $destinationPath));
+
         $this->filesystem->copy(
             $this->configuration->getSpecificationFilePath(),
             $destinationPath
         );
+    }
+
+    private function initWarningPrinting(InputInterface $input): void
+    {
+        $formatter = static function (int $id, string $error, string $file, int $line, array $context) {};
+        if (!$input->getOption('quiet')) {
+            $formatter = new WarningFormatter(
+                new SymfonyStyle(
+                    new ArgvInput(),
+                    new ConsoleOutput()
+                )
+            );
+        }
+
+        set_error_handler($formatter, E_USER_WARNING);
     }
 }