+++ src/Output/Copy/Serializer/ContentType/FormUrlencodedContentTypeSerializer.php
@@ -1,4 +1,5 @@
-<?php declare(strict_types=1);
+<?php
+declare(strict_types=1);
 
 namespace DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType;
 
@@ -7,6 +8,8 @@ use Psr\Http\Message\StreamInterface;
 
 class FormUrlencodedContentTypeSerializer implements ContentTypeSerializerInterface
 {
+    const MIME_TYPE = 'application/x-www-form-urlencoded';
+
     public function encode(SerializableInterface $body): string
     {
         return http_build_query($body->toArray());
@@ -18,5 +21,10 @@ class FormUrlencodedContentTypeSerializer implements ContentTypeSerializerInterf
         parse_str($body->getContents(), $decoded);
 
         return $decoded;
+    }
+
+    public function getMimeType(): string
+    {
+        return self::MIME_TYPE;
     }
 }