+++ src/Output/Copy/Serializer/ContentType/SerializeException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType;
+
+use Exception;
+
+class SerializeException extends Exception
+{
+}