+++ test/suite/functional/Generator/Client/SwaggerPetstoreClient.php
@@ -59,7 +59,9 @@ class SwaggerPetstoreClient
      */
     public function findPets(FindPetsRequest $request): PetCollection
     {
-        return $this->container->get(PetCollectionMapper::class)->toSchema($this->handleResponse($this->sendRequest($request)));
+        $response = $this->handleResponse($this->sendRequest($request));
+
+        return $this->container->get(PetCollectionMapper::class)->toSchema($response);
     }
 
     /**
@@ -69,7 +71,9 @@ class SwaggerPetstoreClient
      */
     public function addPet(AddPetRequest $request): Pet
     {
-        return $this->container->get(PetMapper::class)->toSchema($this->handleResponse($this->sendRequest($request)));
+        $response = $this->handleResponse($this->sendRequest($request));
+
+        return $this->container->get(PetMapper::class)->toSchema($response);
     }
 
     /**
@@ -87,7 +91,9 @@ class SwaggerPetstoreClient
      */
     public function findPetById(FindPetByIdRequest $request): Pet
     {
-        return $this->container->get(PetMapper::class)->toSchema($this->handleResponse($this->sendRequest($request)));
+        $response = $this->handleResponse($this->sendRequest($request));
+
+        return $this->container->get(PetMapper::class)->toSchema($response);
     }
 
     /**